refactor(protocol): split the enrichment tools into registry, context tools, and helpers
